quick fix to limit multiple socket connections

+++ pi_code/capture/capture.js
@@ -29,6 +29,7 @@ var websocket =''
 var browserapp = ''
 
 var hasIPAddress = false;
+var socketcounter =0;
 
 var PIN = [4, 17]; // [SHUTTER, AF]
 var picCt =0; //picture count
@@ -72,45 +73,49 @@ function getIPAddress(cb){
 
 //*** OPEN WEBSOCKET
 function connectSocket(){
-  ws.on('open', function() {
-    http.get("http://localhost:8080/get?eosserialnumber", function(res) {
-      res.on('data', function (chunk) {
-        serialnumber = JSON.parse(chunk.toString())['eosserialnumber']
-        console.log('Camera Serial: '+serialnumber)
-        //console.log('BODY: ' + chunk);
-        var obj = {'address':ipAddress, 'serial':serialnumber}
-        console.log(obj)
-        var objstring = JSON.stringify(obj)
-        ws.send(objstring);
-        newCameraPost(objstring)
+  if(socketCounter < 1){
+    socketCounter++
+    ws.on('open', function() {
+      http.get("http://localhost:8080/get?eosserialnumber", function(res) {
+        res.on('data', function (chunk) {
+          serialnumber = JSON.parse(chunk.toString())['eosserialnumber']
+          console.log('Camera Serial: '+serialnumber)
+          //console.log('BODY: ' + chunk);
+          var obj = {'address':ipAddress, 'serial':serialnumber}
+          console.log(obj)
+          var objstring = JSON.stringify(obj)
+          ws.send(objstring);
+          newCameraPost(objstring)
+        });
+      }).on('error', function(e) {
+        console.log("Got error: " + e.message);
       });
-    }).on('error', function(e) {
-      console.log("Got error: " + e.message);
     });
-  });
 
-  ws.on('close',function(){
-    /** TO DO **/
-    //handle disconnect/attempt reconnect
-    console.log('disconnected')
-  })
+    ws.on('close',function(){
+      /** TO DO **/
+      //handle disconnect/attempt reconnect
+      socketCounter--
+      console.log('disconnected')
+    })
 
 
-  //*** HANDLE WEBSOCKET MESSAGES
-  ws.on('message', function(data, flags) {
-    if(data == 'go'){
-      console.log("trigger shutter, count "+ (picCt++));
-      // flags.binary will be set if a binary data is received
-      // flags.masked will be set if the data was masked
-      hitShutter(); //*** TAKE PICTURE !!
-    }
-    else if(data == 'close'){
-      var child = exec('echo raspberry | sudo shutdown -h now',function(error,stdout,stderr){
-        console.log('stdout: '+stdout)
-        console.log('stderr: '+stderr)
-      })
-    }
-  });
+    //*** HANDLE WEBSOCKET MESSAGES
+    ws.on('message', function(data, flags) {
+      if(data == 'go'){
+        console.log("trigger shutter, count "+ (picCt++));
+        // flags.binary will be set if a binary data is received
+        // flags.masked will be set if the data was masked
+        hitShutter(); //*** TAKE PICTURE !!
+      }
+      else if(data == 'close'){
+        var child = exec('echo raspberry | sudo shutdown -h now',function(error,stdout,stderr){
+          console.log('stdout: '+stdout)
+          console.log('stderr: '+stderr)
+        })
+      }
+    });
+  }
 }
 //**** TAKE A PICTURE !! ****
 function hitShutter(){
